Log Errors when downloading vidx tree
Some time yesterday, an entry was updated to include a URL that does not
resolve to a host. This causes `pack-manager cache pdsc` to stop hard in
it's tracks and refuses to continue past the point at which it stoped.

This PR makes the downloads optional (instead of a result that can fail).
Now `pack-manager cache pdsc` will simply log this an an error including
the url in the message.