cmsis-cli: change app name to "cmsis-pack-manager" to match Python.
If one were to run cmsis-cli and the Python tool (or pyocd) to update
the index, two user data directories would be created. One called
"cmsis-pack", created by cmsis-cli, and the other "cmsis-pack-manager"
created by the Python code. This simple change fixes that so that both
code bases use "cmsis-pack-manager".