Handle inheritance of <debug> element attributes
The spec states: https://open-cmsis-pack.github.io/Open-CMSIS-Pack-Spec/main/html/pdsc_family_pg.html#element_debug

> Multiple debug elements can be defined which are either specific to a
> processor identified by attribute Pname, or which apply to all
> connections.

The current implementation did not handle this correctly: it only
looked for a single debug element defined on either the device or
subfamily. Iif the innermost debug element does not have the
attribute we're looking for, we need to check debug elements on parent
elements as well.

This patch is necessary to correctly parse PSOC 6 CMSIS packs.